launch process with filling
