--- rmcs_ws/src/rmcs_bringup/config/dart-filling-test.yaml
@@ -1,33 +1,0 @@
-rmcs_executor:
-  ros__parameters:
-    update_rate: 1000.0
-    components:
-      - rmcs_core::hardware::CatapultDartFillingTest -> dart_hardware
-      - rmcs_dart_guidance::manager::DartManagerForFillingTest   -> dart_manager
-      - rmcs_dart_guidance::manager::RemoteCommandBridge -> remote_cmd_bridge
-      - rmcs_core::controller::dart::DartFilling -> dart_controller
-
-dart_hardware:
-  ros__parameters:
-    serial_filter: ""
-
-remote_cmd_bridge:
-  ros__parameters:
-    joystick_sensitivity: 0.01
-    
-dart_manager:
-  ros__parameters:
-    # 升降堵转检测参数
-    lifting_stall_threshold:     0.01    # rad/s，低于此值视为堵转
-    lifting_stall_confirm_ticks: 10    # 连续帧数 = 0.1s @ 1kHz
-    lifting_stall_min_run_ticks: 500    # 启动后最少运行帧数，避免启动瞬间误触发
-    lifting_stall_timeout_ticks: 5000   # 超时帧数 = 5s @ 1kHz
-
-    # 限位舵机参数
-    limiting_open_angle:  0xECB    # uint16_t，舵机打开角度（需实测标定）
-    limiting_close_angle: 0xB53   # uint16_t，舵机关闭角度（需实测标定）
-    limiting_fill_ticks:  200    # 填装等待时间 = 0.5s @ 1kHz
-
-dart_controller:
-  ros__parameters:
-    lifting_velocity: 3.0   # rad/s，升降电机速度（正值 = UP时左+右-）
