+++ rmcs_ws/src/rmcs_bringup/config/dart-guidance.yaml
@@ -2,13 +2,14 @@ rmcs_executor:
   ros__parameters:
     update_rate: 1000.0
     components:
-      - rmcs_core::hardware::CatapultDartV4 -> dart_hardware
-      - rmcs_dart_guidance::manager::DartManager -> dart_manager
+      - rmcs_core::hardware::CatapultDartV3Lk        -> dart_hardware
+      - rmcs_dart_guidance::manager::DartManagerV2   -> dart_manager
       - rmcs_dart_guidance::manager::RemoteCommandBridge -> remote_cmd_bridge
       # - rmcs_dart_guidance::manager::WebCommandBridge -> web_cmd_bridge
-      - rmcs_core::controller::dart::DartLaunchSetting -> launch_setting
+      - rmcs_core::controller::dart::DartLaunchSettingV2 -> launch_setting
       - rmcs_core::controller::dart::DartSettingController -> dart_setting
-      - rmcs_core::controller::pid::PidController -> force_screw_velocity_pid_controller
+      - rmcs_core::controller::pid::PidController    -> force_screw_velocity_pid_controller
+      - rmcs_core::broadcaster::ValueBroadcaster -> broadcaster
 
 dart_hardware:
   ros__parameters:
@@ -18,8 +19,15 @@ dart_hardware:
 
 dart_manager:
   ros__parameters:
-    max_transform_rate: 15.0
+    max_transform_rate: 5.0
     manual_force_scale: 5.0
+    limiting_open_angle:  500    # uint16_t，需实测标定
+    limiting_close_angle: 1000   # uint16_t，需实测标定
+    limiting_fill_ticks:  500
+    lifting_stall_threshold:     0.1    # rad/s，低于此值视为堵转
+    lifting_stall_confirm_ticks: 100    # 连续帧数 = 0.1s @ 1kHz
+    lifting_stall_min_run_ticks: 500    # 启动后最少运行帧数，避免启动瞬间误触发
+    lifting_stall_timeout_ticks: 2000   # 超时帧数 = 5s
 
 remote_cmd_bridge:
   ros__parameters:
@@ -27,14 +35,16 @@ remote_cmd_bridge:
 
 launch_setting:
   ros__parameters:
-    belt_velocity: 15.0
+    belt_velocity: 10.0
     sync_coefficient: 0.1
     max_control_torque: 2.5
     trigger_free_value: 0.75
     trigger_lock_value: 0.63
-    b_kp: 1.0
+    b_kp: 5.0
     b_ki: 0.0
     b_kd: 0.0
+    lifting_velocity: 10.0   # rad/s，正值 = 左+右-（平台上升方向），需实测标定方向
+    belt_hold_torque: 15.0    # Nm，传送带下方堵转后的保持扭矩，需实测标定
 
 dart_setting:
   ros__parameters:
@@ -47,9 +57,15 @@ dart_setting:
 
 force_screw_velocity_pid_controller:
   ros__parameters:
-    measurement: /dart/force_screw/velocity
+    measurement: /dart/force_screw_motor/velocity
     setpoint:    /force/control/velocity
-    control:     /dart/force_screw/control_torque
+    control:     /dart/force_screw_motor/control_torque
     kp: 1.0
     ki: 0.0
     kd: 0.0
+
+broadcaster:
+  ros__parameters:
+    forward_list:
+      - /dart/drive_belt/left/control_torque
+      - /dart/drive_belt/right/control_torque
