--- rmcs_ws/src/rmcs_bringup/config/dart-launch-full.yaml
@@ -1,74 +1,0 @@
-rmcs_executor:
-  ros__parameters:
-    update_rate: 1000.0
-    components:
-      - rmcs_core::hardware::CatapultDartV3Lk -> dart_hardware
-      - rmcs_core::controller::dart::DartLaunchControllerV2Full -> launch_controller
-      - rmcs_core::controller::dart::DartSettings -> dart_settings
-      - rmcs_core::controller::pid::PidController -> force_screw_velocity_pid_controller
-      - rmcs_core::controller::pid::PidController -> yaw_velocity_pid_controller
-      - rmcs_core::controller::pid::PidController -> pitch_velocity_pid_controller
-      - rmcs_core::broadcaster::ValueBroadcaster -> broadcaster
-
-broadcaster:
-  ros__parameters:
-    forward_list:
-      - /dart/drive_belt/left/control_torque
-      - /dart/drive_belt/left/velocity
-      - /dart/drive_belt/right/control_torque
-      - /dart/drive_belt/right/velocity
-      - /dart/lifting_left/angle
-      - /dart/lifting_right/angle
-      - /force/sensor/average
-
-dart_hardware:
-  ros__parameters:
-    serial_filter: ""
-    first_sample_spot: 1.0
-    final_sample_spot: 4.0
-
-launch_controller:
-  ros__parameters:
-    belt_velocity: 10.0
-    sync_coefficient: 0.1
-    max_control_torque: 15.0
-    # Trigger servo PWM value 0.0–1.0 (from dart-filling-test calibration)
-    trigger_free_value: 0.75
-    trigger_lock_value: 0.66
-    # Duration to keep trigger free after launch (ms)
-    trigger_free_duration_ms: 300
-    # Time to wait after platform move before belt resets (ms)
-    lifting_settle_ms: 500
-    # Duration to keep limiting servo open for dart feeding (ms)
-    limiting_open_duration_ms: 2000
-    # Lifting motor angles in DEGREES — will be converted to radians in code.
-    # 待标定: 上电后将平台移到 up 位置，记录角度填入 *_up_deg；
-    #         移到 down 位置记录填入 *_down_deg。
-    lifting_up_angle_left_deg:    0.0
-    lifting_down_angle_left_deg:  90.0
-    lifting_up_angle_right_deg:   0.0
-    lifting_down_angle_right_deg: 90.0
-    # Limiting servo (TriggerServo UART, uint16_t hex)
-    limiting_free_angle_: 0x0001
-    limiting_lock_angle_: 0x0050
-
-dart_settings:
-  ros__parameters:
-    is_auto_pitch_control_mode: 0
-    is_auto_force_control_mode: 0
-    max_transform_rate: 5.0
-    pitch_angle_kp: 1.0
-    pitch_angle_ki: 0.0
-    pitch_angle_kd: 0.0
-    force_control_kp: 1.0
-    force_control_ki: 0.0
-    force_control_kd: 0.0
-
-force_screw_velocity_pid_controller:
-  ros__parameters:
-    measurement: /dart/force_screw_motor/velocity
-    setpoint: /force/control/velocity
-    control: /dart/force_screw_motor/control_torque
-    kp: 1.0
-    ki: 0.0
-    kd: 0.0
