--- rmcs_ws/src/rmcs_bringup/config/dart-remot-bridge-test.yaml
@@ -1,16 +1,0 @@
-rmcs_executor:
-  ros__parameters:
-    update_rate: 1000.0
-    components:
-      - rmcs_dart_guidance::manager::RemoteCommandBridge -> remote_cmd_bridge
-      - rmcs_core::hardware::CatapultDartV4 -> dart_hardware
-
-remote_cmd_bridge:
-  ros__parameters:
-    joystick_sensitivity: 0.01
-
-dart_hardware:
-  ros__parameters:
-    serial_filter: ""
-    first_sample_spot: 1.0
-    final_sample_spot: 4.0
